move ops from models to models.architectures
