+++ mmrazor/models/__init__.py
@@ -5,5 +5,4 @@ from .distillers import *  # noqa: F401,F403
 from .losses import *  # noqa: F401,F403
 from .mutables import *  # noqa: F401,F403
 from .mutators import *  # noqa: F401,F403
-from .ops import *  # noqa: F401,F403
 from .task_modules import *  # noqa: F401,F403
