+++ mmrazor/models/architectures/__init__.py
@@ -4,3 +4,4 @@ from .connectors import *  # noqa: F401,F403
 from .dynamic_op import *  # noqa: F401,F403
 from .generators import *  # noqa: F401,F403
 from .heads import *  # noqa: F401,F403
+from .ops import *  # noqa: F401,F403
